Fixing a potential crash in the optimizer.
The GAs could crash at the very end when saving the details of the prediction due to changes in the predictors order. The issue happens when the different predictors are pre-loaded and can differ from one slot to the next. Re-ordering the predictors can lead the Optimizer to access items out of the available array. The solution is to order the predictors after the evaluation.